Updating README.md for CMake (#758)
This is part of the CMake update project #748 

The README.md needs to be updated for CMake. This is my first step at it. Not all changes are done, but this may be a good start.

The standalone instructions are slightly updated, but may not be accurate and will need to be updated later on.

All of the makefile "options" still need to be sorted.